+++ NEWS.md
@@ -4,7 +4,7 @@
 * Begin to flesh out more general conception of CDA
 * `datasets()`: added `ndim` arg; package="vcdExtra" default
 * Added a tidyCat hex logo
-* Generalized vcd::woolf_test() to handle strata better, with special hanlding for 2 x 2 x R x C case. Presently in `extra/`
+* Generalized vcd::woolf_test() to handle strata better, with special handling for 2 x 2 x R x C case. 
 
 ## Version 0.8.7 (2025-12-10)